fixed admin deletion

+++ src/sample/AdminInListController.java
@@ -31,7 +31,7 @@ public class AdminInListController {
     }
 
 
-    public void setMyAccount(Account account) {
+    public void fillData(Account account) {
         myAccount = account;
         usernameLabel.setText(account.getUsername());
         emailLabel.setText(account.getEmail());
@@ -39,37 +39,49 @@ public class AdminInListController {
 
     public void deleteAccount() throws IOException {
 
-        String ownAccountWarning = "";
-        boolean flag = false;
-        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
-
-        if(myAccount.getEmail() == myStore.getActiveAccount().getEmail())
+        if(myStore.getAdminCount() <= 1)
+            deletionNotPossible();
+        else
         {
-            flag = true;
-            ownAccountWarning = "\nThis is your own account so you will be signed out!";
+            String ownAccountWarning = "";
+            boolean flag = false;
+
+            Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+
+            if (myAccount.getEmail() == myStore.getActiveAccount().getEmail()) {
+                flag = true;
+                ownAccountWarning = "\nThis is your own account so you will be signed out!";
+            }
+
+            alert.setTitle("Confirmation");
+
+            alert.setContentText("Are you sure you want to delete this account?" + ownAccountWarning);
+
+            ButtonType buttonTypeYes = new ButtonType("Yes");
+            ButtonType buttonTypeNo = new ButtonType("No");
+
+            alert.getButtonTypes().setAll(buttonTypeYes, buttonTypeNo);
+
+            Optional<ButtonType> result = alert.showAndWait();
+
+            if (result.get() == buttonTypeYes) {
+
+                myStore.deleteAdminAccount(myAccount);
+
+                if (flag)
+                    mainPageController.changeSceneToSignIn();
+                else
+                    mainPageController.changeTabToAdmins();
+            }
         }
 
-        alert.setTitle("Confirmation");
+    }
 
-        alert.setContentText("Are you sure you want to delete this account?" + ownAccountWarning);
-
-        ButtonType buttonTypeYes = new ButtonType("Yes");
-        ButtonType buttonTypeNo = new ButtonType("No");
-
-        alert.getButtonTypes().setAll(buttonTypeYes, buttonTypeNo);
-
-        Optional<ButtonType> result = alert.showAndWait();
-
-        if(result.get() == buttonTypeYes)
-        {
-
-            myStore.deleteAdminAccount(myAccount);
-            if(flag)
-                mainPageController.changeSceneToSignIn();
-            else
-                mainPageController.changeTabToAdmins();
-        }
-
+    private void deletionNotPossible() {
+        Alert alert = new Alert(Alert.AlertType.ERROR);
+        alert.setTitle("Deletion not possible!");
+        alert.setContentText("This is the last admin account!");
+        alert.showAndWait();
     }
 
 }
